Fix JavaScript not executing
This was due to the Gemfile.lock file having a reference to an older
version of Webpacker (pre.2) that would have required us to use the
javascript_packs_with_chunks_tag tag for things to work.

But upgrading to beta.7 fixes things so we can continue using the
existing javascript_pack_tag.