Don't delete old assets by default
There is a slight blip between deleting the old assets and copying the
new ones where if someone were to make a request in between this period
of time they would get no assets served.

The odds of this happening are extremely low since these operations are
nearly instant but still as a best practice if you want to clean up
older precompiled assets it's worth doing it in a cron job.

With cp -r you can use the mtime of the manifest JSON file to delete
precompiled files that have an older mtime.